+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-07-19
+
+### Added
+- Add jest preset
+
+```js
+// jest.config.js on your project
+const jestPreset = require('rocket-punch/jest-preset');
+
+module.exports = {
+  ...jestPreset,
+  
+  // customize your config
+  collectCoverageFrom: [
+    ...jestPreset.collectCoverageFrom,
+    '!src/@rocket-scripts/*/commands.ts',
+    '!src/@rocket-scripts/cli/**',
+  ],
+};
+```
+
 ## [1.0.0] - 2020-07-03
 
 Release
@@ -23,6 +44,7 @@ Release
 ### Added
 - Rewrite paths by TypeScript compiler transformer API on build (e.g. `import A from '../a'` to `import A from 'a'`)
 
+[1.1.0]: https://github.com/rocket-hangar/rocket-punch/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/rocket-hangar/rocket-punch/compare/v1.0.0-alpha.9...v1.0.0
 [1.0.0-alpha.9]: https://github.com/rocket-hangar/rocket-punch/compare/v1.0.0-alpha.5...v1.0.0-alpha.9
 [1.0.0-alpha.5]: https://github.com/rocket-hangar/rocket-punch/releases/tag/v1.0.0-alpha.5